Rename instruction-executor.ts
